fix(pg-delta): classify zero-row autoSeed inserts and rethrow seed-coverage errors past EXPECTED_RED pins
Two confirmed review findings on #350.

Finding 1 — `seeded` false positive on trigger-suppressed inserts.
A resolved INSERT ... DEFAULT VALUES does not guarantee a stored row: a
BEFORE INSERT trigger returning NULL (or a DO INSTEAD rule) makes the
statement SUCCEED with rowCount 0. autoSeedEmptyTables now inspects the
pg driver's rowCount: >= 1 -> `seeded`; 0 -> `skipped` with the synthetic
reasonCode "no_row" (documented as NOT a SQLSTATE on SeedOutcome and in
the allowlist header), routed through the same allowlist gate — so the
trigger scenario now needs a conscious allowlist entry.
RED (tests/autoseed-outcomes.test.ts):
  expect(outcomeFor("trigger_suppressed")).toEqual({ status: "skipped",
    reasonCode: "no_row" }) -> received { status: "seeded" }

Finding 2 — EXPECTED_RED swallowed seed-coverage violations.
enforceSeedCoverage threw a plain Error, and the pinned driver path is
`try { runDirection() } catch { return }`, so a seed-coverage regression
inside a pinned scenario satisfied the pin. Introduced SeedCoverageError
and moved SeedCoverageError + enforceSeedCoverage into an importable
tests/seed-coverage.ts; the pinned catch now re-throws SeedCoverageError.
RED (tests/seed-coverage.test.ts):
  import { enforceSeedCoverage, SeedCoverageError } from "./seed-coverage.ts"
    -> error: Cannot find module './seed-coverage.ts'

Corpus (postgres:17-alpine): 632 pass / 0 fail. Added 1 allowlist entry
(trigger-operations--trigger-drop-before-function-drop forward,
test_schema.foo, reasonCode "no_row"); allowlist now 170 (169x 23502 + 1x
no_row).
