+++ packages/pg-delta/src/proof/prove.ts
@@ -33,10 +33,13 @@ export interface TableRef {
  *  `ProveOptions.autoSeed`). Taxonomy is by SQLSTATE class, NOT string matching:
  *  - `seeded` — a synthetic `DEFAULT VALUES` row landed; the table now has
  *    content-fingerprint coverage in the data-preservation check.
- *  - `skipped` — the insert hit a class-23 integrity-constraint violation
- *    (`reasonCode` = the SQLSTATE: `23502` NOT NULL w/o default, `23503` FK,
- *    `23505` unique, `23514` check, any `23xxx`). Genuinely unseedable with
- *    defaults — expected, not a defect; the table keeps `contentMode: "none"`.
+ *  - `skipped` — no row was stored and that is expected, in two shapes:
+ *    (a) the insert hit a class-23 integrity-constraint violation (`reasonCode`
+ *    = the SQLSTATE: `23502` NOT NULL w/o default, `23503` FK, `23505` unique,
+ *    `23514` check, any `23xxx`); or (b) the insert SUCCEEDED but stored zero
+ *    rows because a BEFORE INSERT trigger returned NULL / a DO INSTEAD rule
+ *    suppressed it (`reasonCode` = the synthetic sentinel `"no_row"`, which is
+ *    NOT a SQLSTATE). Either way the table keeps `contentMode: "none"`.
  *  - `failed` — anything else (a raised exception, connection/syntax/permission
  *    error, or a driver error with no code): a real problem the caller must see
  *    rather than have swallowed. `reasonCode` is the SQLSTATE when the driver
@@ -283,10 +286,20 @@ async function autoSeedEmptyTables(
     // raised exception, connection/syntax/permission error, unknown) is a real
     // problem → `failed`, so it can't hide behind `contentMode: "none"`.
     try {
-      await pool.query(
+      const res = await pool.query(
         `INSERT INTO ${qte(schema)}.${qte(name)} DEFAULT VALUES`,
       );
-      outcomes.push({ table: ref, status: "seeded" });
+      // a resolved insert is NOT proof of a stored row: a BEFORE INSERT trigger
+      // that RETURNS NULL (or a DO INSTEAD NOTHING rule) makes the statement
+      // SUCCEED with rowCount 0 — nothing was seeded. Classify by rowCount so
+      // that case becomes `skipped` (synthetic reasonCode "no_row", NOT a
+      // SQLSTATE) and goes through the same allowlist gate, instead of a false
+      // `seeded` that leaves the table at contentMode "none" with no coverage.
+      if ((res.rowCount ?? 0) >= 1) {
+        outcomes.push({ table: ref, status: "seeded" });
+      } else {
+        outcomes.push({ table: ref, status: "skipped", reasonCode: "no_row" });
+      }
     } catch (err) {
       const rawCode = (err as { code?: unknown }).code;
       const code = typeof rawCode === "string" ? rawCode : undefined;
