+++ packages/pg-delta/tests/engine.test.ts
@@ -14,8 +14,8 @@ import { extract } from "../src/extract/extract.ts";
 import { plan } from "../src/plan/plan.ts";
 import { probeApplierCapability } from "../src/policy/capability.ts";
 import { rel } from "../src/plan/render.ts";
-import { provePlan, type ProofVerdict } from "../src/proof/prove.ts";
-import { isSeedSkipAllowed } from "./autoseed-allowlist.ts";
+import { provePlan } from "../src/proof/prove.ts";
+import { enforceSeedCoverage, SeedCoverageError } from "./seed-coverage.ts";
 import { loadCorpus, type Scenario } from "./corpus.ts";
 import {
   isolatedClusterPair,
@@ -23,66 +23,6 @@ import {
   type Cluster,
 } from "./containers.ts";
 import { EXPECTED_RED } from "./expected-red.ts";
-
-/**
- * Enforce the corpus auto-seed coverage contract on a proof verdict.
- *  - a `failed` seed outcome (anything NOT class-23 — a raised exception,
- *    connection/permission error, …) fails the scenario loudly; it is never
- *    allowlistable.
- *  - a class-23 `skipped` outcome is tolerated ONLY when its precise
- *    { scenario, direction, table, reasonCode } key is in the checked-in
- *    allowlist (tests/autoseed-allowlist.ts); an unlisted skip fails the
- *    scenario so newly-lost data-preservation coverage can't slip in silently.
- * A `SEED_AUDIT {json}` line is written to fd 2 before failing so a full corpus
- * run yields ready-to-paste allowlist keys.
- */
-function enforceSeedCoverage(
-  scenarioName: string,
-  direction: "forward" | "reverse",
-  label: string,
-  verdict: ProofVerdict,
-): void {
-  const outcomes = verdict.seedOutcomes ?? [];
-  if (outcomes.length === 0) return;
-  const coverageMode = (t: { schema: string; name: string }): string =>
-    verdict.coverage.perTable.find(
-      (p) => p.table.schema === t.schema && p.table.name === t.name,
-    )?.contentMode ?? "none";
-  const violations: string[] = [];
-  for (const o of outcomes) {
-    if (o.status === "seeded") continue;
-    if (o.status === "failed") {
-      writeSync(
-        2,
-        `SEED_AUDIT ${JSON.stringify({ status: "failed", scenario: scenarioName, direction, table: o.table, reasonCode: o.reasonCode ?? null })}\n`,
-      );
-      violations.push(
-        `  FAILED seed ${rel(o.table.schema, o.table.name)} ` +
-          `(coverage=${coverageMode(o.table)}, sqlstate=${o.reasonCode ?? "none"}): ${o.message}`,
-      );
-      continue;
-    }
-    // class-23 skip: tolerated only if declared in the allowlist
-    if (isSeedSkipAllowed(scenarioName, direction, o.table, o.reasonCode)) {
-      continue;
-    }
-    writeSync(
-      2,
-      `SEED_AUDIT ${JSON.stringify({ status: "skipped", scenario: scenarioName, direction, table: o.table, reasonCode: o.reasonCode })}\n`,
-    );
-    violations.push(
-      `  UNLISTED class-23 skip ${rel(o.table.schema, o.table.name)} ` +
-        `(coverage=${coverageMode(o.table)}, sqlstate=${o.reasonCode}) — ` +
-        `add to tests/autoseed-allowlist.ts if genuinely unseedable`,
-    );
-  }
-  if (violations.length > 0) {
-    throw new Error(
-      `[${label}] autoSeed coverage contract violated ` +
-        `(scenario=${scenarioName}, direction=${direction}):\n${violations.join("\n")}`,
-    );
-  }
-}
 
 async function proveOn(
   name: string,
@@ -268,7 +208,10 @@ async function runPinnedOrProve(
   }
   try {
     await runDirection(scenario, direction);
-  } catch {
+  } catch (error) {
+    // a seed-coverage violation is NEVER a legitimate "red as pinned" — it must
+    // fail the corpus even inside a pinned scenario, so re-throw it.
+    if (error instanceof SeedCoverageError) throw error;
     return; // red as pinned — fine
   }
   throw new Error(
